+++ context/context.go
@@ -364,6 +364,8 @@ type Context interface {
 	//  | Various Request and Post Data                              |
 	//  +------------------------------------------------------------+
 
+	// URLParam returns true if the url parameter exists, otherwise false.
+	URLParamExists(name string) bool
 	// URLParamDefault returns the get parameter from a request, if not found then "def" is returned.
 	URLParamDefault(name string, def string) string
 	// URLParam returns the get parameter from a request , if any.
@@ -745,99 +747,6 @@ type Context interface {
 	// to be executed at serve-time. The full app's fields
 	// and methods are not available here for the developer's safety.
 	Application() Application
-
-	//  +--------------------------------------------------------------+
-	//  | https://github.com/golang/net/blob/master/context/context.go |                                     |
-	//  +--------------------------------------------------------------+
-
-	// Deadline returns the time when work done on behalf of this context
-	// should be canceled.  Deadline returns ok==false when no deadline is
-	// set.  Successive calls to Deadline return the same results.
-	Deadline() (deadline time.Time, ok bool)
-
-	// Done returns a channel that's closed when work done on behalf of this
-	// context should be canceled.  Done may return nil if this context can
-	// never be canceled.  Successive calls to Done return the same value.
-	//
-	// WithCancel arranges for Done to be closed when cancel is called;
-	// WithDeadline arranges for Done to be closed when the deadline
-	// expires; WithTimeout arranges for Done to be closed when the timeout
-	// elapses.
-	//
-	// Done is provided for use in select statements:
-	//
-	//  // Stream generates values with DoSomething and sends them to out
-	//  // until DoSomething returns an error or ctx.Done is closed.
-	//  func Stream(ctx context.Context, out chan<- Value) error {
-	//  	for {
-	//  		v, err := DoSomething(ctx)
-	//  		if err != nil {
-	//  			return err
-	//  		}
-	//  		select {
-	//  		case <-ctx.Done():
-	//  			return ctx.Err()
-	//  		case out <- v:
-	//  		}
-	//  	}
-	//  }
-	//
-	// See http://blog.golang.org/pipelines for more examples of how to use
-	// a Done channel for cancelation.
-	Done() <-chan struct{}
-
-	// Err returns a non-nil error value after Done is closed.  Err returns
-	// Canceled if the context was canceled or DeadlineExceeded if the
-	// context's deadline passed.  No other values for Err are defined.
-	// After Done is closed, successive calls to Err return the same value.
-	Err() error
-
-	// Value returns the value associated with this context for key, or nil
-	// if no value is associated with key.  Successive calls to Value with
-	// the same key returns the same result.
-	//
-	// Use context values only for request-scoped data that transits
-	// processes and API boundaries, not for passing optional parameters to
-	// functions.
-	//
-	// A key identifies a specific value in a Context.  Functions that wish
-	// to store values in Context typically allocate a key in a global
-	// variable then use that key as the argument to context.WithValue and
-	// Context.Value.  A key can be any type that supports equality;
-	// packages should define keys as an unexported type to avoid
-	// collisions.
-	//
-	// Packages that define a Context key should provide type-safe accessors
-	// for the values stores using that key:
-	//
-	// 	// Package user defines a User type that's stored in Contexts.
-	// 	package user
-	//
-	// 	import "golang.org/x/net/context"
-	//
-	// 	// User is the type of value stored in the Contexts.
-	// 	type User struct {...}
-	//
-	// 	// key is an unexported type for keys defined in this package.
-	// 	// This prevents collisions with keys defined in other packages.
-	// 	type key int
-	//
-	// 	// userKey is the key for user.User values in Contexts.  It is
-	// 	// unexported; clients use user.NewContext and user.FromContext
-	// 	// instead of using this key directly.
-	// 	var userKey key = 0
-	//
-	// 	// NewContext returns a new Context that carries value u.
-	// 	func NewContext(ctx context.Context, u *User) context.Context {
-	// 		return context.WithValue(ctx, userKey, u)
-	// 	}
-	//
-	// 	// FromContext returns the User value stored in ctx, if any.
-	// 	func FromContext(ctx context.Context) (*User, bool) {
-	// 		u, ok := ctx.Value(userKey).(*User)
-	// 		return u, ok
-	// 	}
-	Value(key interface{}) interface{}
 }
 
 // Next calls all the next handler from the handlers chain,
@@ -1329,6 +1238,7 @@ func (ctx *context) ContentType(cType string) {
 			cType += "; charset=" + charset
 		}
 	}
+
 	ctx.writer.Header().Set(contentTypeHeaderKey, cType)
 }
 
@@ -1365,6 +1275,16 @@ func (ctx *context) GetStatusCode() int {
 //  +------------------------------------------------------------+
 //  | Various Request and Post Data                              |
 //  +------------------------------------------------------------+
+
+// URLParam returns true if the url parameter exists, otherwise false.
+func (ctx *context) URLParamExists(name string) bool {
+	if q := ctx.request.URL.Query(); q != nil {
+		_, exists := q[name]
+		return exists
+	}
+
+	return false
+}
 
 // URLParamDefault returns the get parameter from a request, if not found then "def" is returned.
 func (ctx *context) URLParamDefault(name string, def string) string {
@@ -2397,15 +2317,10 @@ func (ctx *context) ServeFile(filename string, gzipCompression bool) error {
 	defer f.Close()
 	fi, _ := f.Stat()
 	if fi.IsDir() {
-		filename = path.Join(filename, "index.html")
-		f, err = os.Open(filename)
-		if err != nil {
-			return fmt.Errorf("%d", 404)
-		}
-		fi, _ = f.Stat()
+		return ctx.ServeFile(path.Join(filename, "index.html"), gzipCompression)
 	}
-	return ctx.ServeContent(f, fi.Name(), fi.ModTime(), gzipCompression)
 
+	return ctx.ServeContent(f, fi.Name(), fi.ModTime(), gzipCompression)
 }
 
 // SendFile sends file for force-download to the client
@@ -2672,111 +2587,4 @@ func (ctx *context) Exec(method string, path string) {
 // and methods are not available here for the developer's safety.
 func (ctx *context) Application() Application {
 	return ctx.app
-}
-
-//  +--------------------------------------------------------------+
-//  | https://github.com/golang/net/blob/master/context/context.go |                                     |
-//  +--------------------------------------------------------------+
-
-// Deadline returns the time when work done on behalf of this context
-// should be canceled.  Deadline returns ok==false when no deadline is
-// set.  Successive calls to Deadline return the same results.
-func (ctx *context) Deadline() (deadline time.Time, ok bool) {
-	return
-}
-
-// Done returns a channel that's closed when work done on behalf of this
-// context should be canceled.  Done may return nil if this context can
-// never be canceled.  Successive calls to Done return the same value.
-//
-// WithCancel arranges for Done to be closed when cancel is called;
-// WithDeadline arranges for Done to be closed when the deadline
-// expires; WithTimeout arranges for Done to be closed when the timeout
-// elapses.
-//
-// Done is provided for use in select statements:
-//
-//  // Stream generates values with DoSomething and sends them to out
-//  // until DoSomething returns an error or ctx.Done is closed.
-//  func Stream(ctx context.Context, out chan<- Value) error {
-//  	for {
-//  		v, err := DoSomething(ctx)
-//  		if err != nil {
-//  			return err
-//  		}
-//  		select {
-//  		case <-ctx.Done():
-//  			return ctx.Err()
-//  		case out <- v:
-//  		}
-//  	}
-//  }
-//
-// See http://blog.golang.org/pipelines for more examples of how to use
-// a Done channel for cancelation.
-func (ctx *context) Done() <-chan struct{} {
-	return nil
-}
-
-// Err returns a non-nil error value after Done is closed.  Err returns
-// Canceled if the context was canceled or DeadlineExceeded if the
-// context's deadline passed.  No other values for Err are defined.
-// After Done is closed, successive calls to Err return the same value.
-func (ctx *context) Err() error {
-	return nil
-}
-
-// Value returns the value associated with this context for key, or nil
-// if no value is associated with key.  Successive calls to Value with
-// the same key returns the same result.
-//
-// Use context values only for request-scoped data that transits
-// processes and API boundaries, not for passing optional parameters to
-// functions.
-//
-// A key indentifies a specific value in a context.  Functions that wish
-// to store values in context typically allocate a key in a global
-// variable then use that key as the argument to context.WithValue and
-// context.Value.  A key can be any type that supports equality;
-// packages should define keys as an unexported type to avoid
-// collisions.
-//
-// Packages that define a context key should provide type-safe accessors
-// for the values stores using that key:
-//
-// 	// Package user defines a User type that's stored in Contexts.
-// 	package user
-//
-// 	import "golang.org/x/net/context"
-//
-// 	// User is the type of value stored in the Contexts.
-// 	type User struct {...}
-//
-// 	// key is an unexported type for keys defined in this package.
-// 	// This prevents collisions with keys defined in other packages.
-// 	type key int
-//
-// 	// userKey is the key for user.User values in Contexts.  It is
-// 	// unexported; clients use user.NewContext and user.FromContext
-// 	// instead of using this key directly.
-// 	var userKey key = 0
-//
-// 	// NewContext returns a new context that carries value u.
-// 	func NewContext(ctx context.Context, u *User) context.Context {
-// 		return context.WithValue(ctx, userKey, u)
-// 	}
-//
-// 	// FromContext returns the User value stored in ctx, if any.
-// 	func FromContext(ctx context.Context) (*User, bool) {
-// 		u, ok := ctx.Value(userKey).(*User)
-// 		return u, ok
-// 	}
-func (ctx *context) Value(key interface{}) interface{} {
-	if key == 0 {
-		return ctx.request
-	}
-	if k, ok := key.(string); ok {
-		return ctx.values.GetString(k)
-	}
-	return nil
 }