+++ doc.go
@@ -35,7 +35,7 @@ Source code and other details for the project are available at GitHub:
 
 Current Version
 
-8.4.2
+8.4.4
 
 Installation
 
@@ -110,6 +110,12 @@ Example code:
         // Method GET: http://localhost:8080/profile/anytypeofstring
         app.Get("/profile/{username:string}", profileByUsername)
 
+        // Want to use a custom regex expression instead?
+        // Easy: app.Get("/profile/{username:string regexp(^[a-zA-Z ]+$)}")
+        //
+        // If parameter type is missing then it's string which accepts anything,
+        // i.e: /{paramname} it's exactly the same as /{paramname:string}.
+
         usersRoutes := app.Party("/users", logThisMiddleware)
         {
             // Method GET: http://localhost:8080/users/42
@@ -173,7 +179,7 @@ Below you'll see some useful examples:
     app.Run(iris.Addr(":8080"))
 
 
-    // Same as before but using a custom http.Server which may being used somewhere else too
+    // Same as before but using a custom http.Server which may be in use somewhere else too
     app.Run(iris.Server(&http.Server{Addr:":8080"}))